+++ exercises/matching-brackets/.meta/solutions/matching_brackets.rb
